26012025_AG_AdminUserList
26012025_AG_AdminUserList

+++ app/api/cakes/[id]/route.ts
@@ -4,6 +4,8 @@ import { Cake } from '@/models/cake';
 import { getServerSession } from 'next-auth/next';
 import { authOptions } from "@/lib/auth";
 import { ObjectId } from 'mongodb';
+import { deleteFromCache } from '@/lib/redis';
+//import { AdminSettings } from '@/models/adminSettings';
 export async function GET(request: Request, { params }: { params: { id: string } }) {
   try {
     await clientPromise;
@@ -45,6 +47,10 @@ export async function PUT(request: Request, { params }: { params: { id: string }
     if (!cake) {
       return NextResponse.json({ error: 'Cake not found' }, { status: 404 });
     }
+   // Delete the cache for cakes to ensure fresh data
+    const cacheKey = 'cakes';
+    await deleteFromCache(cacheKey);
+
     return NextResponse.json(cake);
   } catch  {
     return NextResponse.json({ error: 'Failed to update cake' }, { status: 500 });
@@ -53,11 +59,21 @@ export async function PUT(request: Request, { params }: { params: { id: string }
 
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
   try {
+    
+    const session = await getServerSession(authOptions);
+    if (!session || session.user.role !== 'admin') {
+      return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+    }
+  
+    //console.log(params.id,"delte")
     await clientPromise;
     const cake = await Cake.findByIdAndDelete(params.id);
     if (!cake) {
       return NextResponse.json({ error: 'Cake not found' }, { status: 404 });
     }
+     // Delete the cache for cakes to ensure fresh data
+     const cacheKey = 'cakes';
+     await deleteFromCache(cacheKey);
     return NextResponse.json({ message: 'Cake deleted successfully' });
   } catch  {
     return NextResponse.json({ error: 'Failed to delete cake' }, { status: 500 });
